requirements.txt
streamlit
streamlit-drawable-canvas
numpy
opencv-python
Pillow
onnxruntime
requests